requirements.txt
# Anthropic Claude AI API for content generation
anthropic>=0.39.0

# Google Sheets integration
gspread>=5.12.0
oauth2client>=4.1.3

# Slack notifications
slack-sdk>=3.26.0

# Environment variable management
python-dotenv>=1.0.0

# Date/time handling
pytz>=2024.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0
